#41 Split Identity model into files
