+++ src/langium-model-server/identity/model/error.ts
