+++ src/langium-model-server/identity/model/name.ts
@@ -1,0 +1,4 @@
+
+export type DerivativeIdentityName = object
+export type AstNodeIdentityName = string
+export type IdentityName = AstNodeIdentityName | DerivativeIdentityName
