+++ src/langium-model-server/identity/model/semantic-identifier.ts
@@ -1,0 +1,11 @@
+import * as uuid from 'uuid'
+
+export type SemanticIdentifier = {
+    id: string
+}
+
+export namespace SemanticIdentifier {
+    export function generate(): string {
+        return uuid.v4()
+    }
+}
